requirements.txt
# MOSEE - Margin of Safety & Earnings to Equity Analyzer
# Required dependencies

# Data retrieval
yfinance>=0.2.54  # Critical: fixes for Yahoo's 2024 API changes
forex-python>=1.8

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization (for PDF reports)
matplotlib>=3.7.0

# PDF generation
reportlab>=4.0.0

# Database (PostgreSQL - Vercel Postgres)
psycopg2-binary>=2.9.0

# yfinance dependencies
lxml>=4.9.0
beautifulsoup4>=4.12.0
html5lib>=1.1
appdirs>=1.4.4
frozendict>=2.3.0
peewee>=3.16.0
multitasking>=0.0.11

# Utilities
python-dotenv>=1.0.0  # For loading environment variables

# AI Analysis (annual report analysis with Gemini)
google-generativeai>=0.8.0  # Gemini API client
chromadb>=0.5.0             # Local vector database for RAG knowledge base
sentence-transformers>=3.0.0  # Embedding model (all-MiniLM-L6-v2)
edgartools>=3.0.0           # SEC EDGAR filing fetcher with section extraction

# Optional: for local development/testing
rich>=13.0.0  # Pretty console output for debugging
